Update patch set 4
Patch Set 4:

Thanks for your review. I've put a comment on the EAX.
I requested EAX as a temp because I think that will minimize the generated code. We need to put the original lock word into EAX before the cmpxchg. The most efficient way is directly allocate EAX as a temp and load the lock word into that temp register. For x86, by default the object address will be loaded to EAX (calling_convention.GetRegisterAt(0) will return EAX in x86). If we did not change that, we need to move the object from EAX to anoter temp and move the lock word into EAX before cmpxchg and move the obj back to EAX in slow path. This will introduce addtional move instructions. So, I think the most efficient way is to allocate the obj to another register and directly declair a temp register to be EAX for cmpxchg.
For x86_64 things are better, because in the x86_64 the obj will be put into RDI which does not confilct with cmpxchg and we do not need an additional move instruction in the slow path.

Patch-set: 4
Label: Verified=0