Update patch set 2
Patch Set 2:

> Abandoned
 > 
 > Calin wants to do this his way


you probably want to revive this and add the statuses you seem to want here :).

What I needed is you to wait merging this until the other change is in so I don't have to deal with the conflicts and the potential "side-effects".

Patch-set: 2
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0