Update patch set 2
Patch Set 2:

Hmm, then Nicolas may be right that the code is not being used now, though the intention was for it to be. I'll try to look into why it's not being generated later, but this is a fix I'd like to go in.

Patch-set: 2
Reviewer: Gerrit User 1010108 <1010108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0