Update patch set 1
Patch Set 1:

> > Do we have a bug open for this issue with 'ps'?
 > 
 > That's on M, using our own implementation. We are switching to
 > toybox, so I don't think we need a bug.

a bug would be useful, though, because (a) we're not switching to toybox ps for N and (b) this is the first i've heard of it, so i'm a bit suspicious that you're the only person having trouble with one of the most heavily-used commands. do you have an strace of a hung ps?

Patch-set: 1