Update patch set 1
Abandoned

I have measured the performance impact of this and it was worse then I expected.  Facebook start up time increased from 2s to 10s.

I think I will go with the opposite approach for the time being - use of the AOT code LLDB and make the best of the stackmap that are there.

I might reconsider this in the future or control it with flag.

Patch-set: 1
Status: abandoned