Update patch set 1
Patch Set 1: Verified+1

> > IIUC, this header inclusion is needed by an instruction in
 > > SwitchTable's ctor on line 6495:
 > >
 > > int32_t table_offset = instruction.VRegB_31t();
 > >
 > > What about passing this `table_offset` value as an argument to
 > the
 > > SwitchTable's ctor from the Builder instead?
 > 
 > Please don't complicate things more than necessary. I'm moving the
 > class to a separate header file in my CL.

Agree with David. Point of the class is to encapsulate all the reading of the instruction and data.

(David, I need this CL, as otherwise kIsDebugBuild=true for a libart build fails, but I need that for FB. As your CL seems still WIP, I'll merge this, and I think rebase should be trivial for you...)

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1