Update patch set 4
Patch Set 4:

We also convinced ourselves that it's not necessary to check the sign; the code for the positive case should always work.  Reasoning from https://buganizer.corp.google.com/issues/26327751 :

I think it can possibly fail only in [-1,0], since otherwise again |x - result| < x.  For x in [-1,1/2], the answer should still be exact, so we're fine.  For -1/2 + 1ulp or larger, we get rounding, but I think the difference should be -1/2 or larger, which should still yield the correct result.  Near zero, with a floor of -1, I could get a difference of exactly 1, but that still produces the correct result.

Ask me if you want the Java test program that somewhat confirms this experimentally.

We may have to be careful to get NaN handling right in assembly code.

Patch-set: 4