Fix memory leak caused by optimization after strict-aliasing
list<_Tp, _Alloc>::remove will create temporary object that collects
(splices) deleted nodes. It is assumed that memory allocated for this
object is released when it goes out of scope. It was discovered that
inlined destructor of __list_imp class (base of list) is silently
optimized and __list_imp::clear is removed as latter breaks TBAA rules
due to __list_imp::_end_ declared as __list_node_base then casted to
__list_node derived from that.

There is known issue for that with GCC
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=61571 that recommends
use -fno-strict-aliasing for the source code but this can cause
performance degrades for art runtime.

Fix proposal is to use list<_Tp, _Alloc>::remove_if that does not
allocate temporary objects thus avoids creating optmized code.

Change-Id: I47e93520d3d1d65a1df42259e0cf55c8690150e9