Update patch set 1
Patch Set 1:

It is not a show-stopper, but the blocking JIT has bigger cost then I thought.  Enough to make me a bit uncomfortable doing it by default.

The average developer does not need to debug framework code, so as long as I have a story for showing backtrace though framework code, it is good enough for me.  I think there no reliable way for them to get the source code for framework in the first place.

On the other hand, I should probably keep this and just guard it under some flag. I would like to have flow for internal people who develop mixed C++/Java framework code as well.  On the other hand, better solution might be for them to just compile boot.oat as --native-debuggable.

Patch-set: 1