Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > I would like to also keep the whole thing in oatdump.  I agree that
 > separate tool would be better, but let's stay consistent. The
 > command "oatdump --symbolize=boot.oat --output=boot.oat.symbols"
 > does work in M and we use it in LLDB.
 > 
 > --symbolize-with-debug-info is misleading.  --symbolize includes
 > debug-info as well.  Both versions make a copy of ELF file and they
 > add symbols (including DWARF data now). The second version only
 > differs in that it omits .text, and .rodata. Because why copy the
 > whole bulk of boot.oat if you just care about the symbols.
 > 
 > What about oatdump --symbolize=boot.oat --output=boot.oat.symbols
 > --only-keep-debug? That is, the only difference is the
 > --only-keep-debug flag which tells us to omit .text/.rodata (name
 > taken from "strip" command with identical semantic).

A single tool is easier to argue for (and expect to find it on non-Nexus devices)...

Also, the profiling people would likely like to not have to make symbolization dependent on the Android version. Does anyone have problems to unconditionally do this? The symbolization was meant originally to make both debugging and profiling a bit easier, so why not keep it as a single thing?

(And yes, "symbolize" is then a bit misleading, so maybe add an alias, so we have a good name, and backwards compatibility?)

Patch-set: 4