Update patch set 5
Patch Set 5:

Great to see you chiming in Ian! 

> (5 comments)
 > 
 > this change goes from O(N) iftable search to an O(N) search of a
 > conflict array, with associated memory overhead. Why not JIT
 > conflict handlers? ie in the conflict handler JIT a routine that
 > does:
 > if (receiver.class == ...) { call ...}
 > else { call slow path}

That's probably what I'd ideally want, but we lack the infrastructure for now in building such stubs for the JIT. Probably in the near future.

You mention memory overhead of the conflict table, but the JIT stub is probably in the same range of memory overhead, don't you think?

Patch-set: 5