Update patch set 1
Patch Set 1: Code-Review-1

> This CL is mostly for my sanity, so that we have one source of
 > truth.
 > Before the CL the behavior changes depending on whether and when we
 > attach Java debugger.
 > With this CL the behavior is at least consistent purely based on
 > "am start" flags.
 > 
 > Based on Oleskiy's measurements from today, it appears we might
 > need to make the compilation optional (and probably off by default)
 > for native debugging.  Which makes it even more desirable to have
 > one place to control it.
 > 
 > I am open to suggestions where the code should be placed, and how
 > the potential system property to control the boot.oat JIT during
 > native debugging should be called.

I don't like that approach. Here, every app that allows debugging will pay the performance penalty of switching boot image to interpreter (and JIT) instead of running the AOT-compiled code. In userdebug and eng, this means almost *every* app will pay this price. This would not be active only in user build then (where the debuggability only comes from the AndroidManifest's android:debug flag), not very productive for development purposes.

The goal of the original CL was to pay that price *only* when a debugger is connected, not when it's possible a debugger may attach in the future. I find it pretty dangerous to do this constantly.

Patch-set: 1
Label: Code-Review=-1