Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > I am not sure why this is necessary. AFAIK I didn't change any of
 > the logic around interface fields with my default methods work.

Where did I write anything about default methods? Seems I must have typed that in my sleep...

This is not about your work. You're a reviewer because you know this code, and you know interfaces.

This is necessary because there's still a significant number of broken apps in the app store, and there's no hope of fixing them all. 

You are *partially* right about the abstract-ness issue. We did *not* return false here. We rejected the class with a VerifyError in the verifier. The dex file wasn't rejected outright, so this *is* the right change. (Except I should the kIsDebugBuild in the method verifier...)

Patch-set: 1