Update patch set 1
Patch Set 1:

> > Should we split this test?
 > 
 > Would that really help? I'd rather reduce the size of the boot
 > image, especially the number of stack maps we are dumping with this
 > test.

I was going for the quick and simple solution: if this test is doing too much, maybe the workload should be distributed over several tests (which would give them a chance not to be killed by the timeout).  Improving granularity is also better with respect to `make -j N`.

Patch-set: 1