Update patch set 4
Patch Set 4:

> > Is that the right solution? Is a merge of A & error = A? That
 > > should be documented with reasoning...
 > 
 > I don't know, which is why this was a WIP-discussion CL. If merge
 > of A & invalid = A, this seemed just as good a guess as any (and it
 > actually resolved the crash). I was hoping for some guidance from
 > the LON experts, like David?

Sorry for the late comment. I missed the notification on it.
I am not sure what IsErrornous actually imply but it does'n look like the right solution.

The safe way here is to  fall back to Object if any of the type is errornous. 

Also, as part of the fix we should make sure that IsObjectClass() does not return true if it's not really object (and only errornous)

Patch-set: 4