Update patch set 3
Patch Set 3:

Thanks for the fix Alex.

It looks to me we are (as well as in the bug) lucky to hit a conflicting entry and then your fix is resolving it properly at runtime. What happens if we don't hit a conflicting entry? Can we actually invoke an unrelated method?

Patch-set: 3