Update patch set 1
Patch Set 1:

> > > (Personally I think having the runtime parse a compiler option
 > is
 > > > yucky.)
 > >
 > > I thought that's what you and Richard wanted? What are you
 > > suggesting then?
 > 
 > the thought was to have a proper runtime option and stop pretending
 > the runtime doesn't know about such things.

So this CL changes the runtime to stop pretending it doesn't know the meaning of the -Xcompiler-option argument, thus making -Xcompiler-option a proper runtime option.

I would have preferred to have the runtime parse all the compiler options and store the parsed compiler options instead of the raw string, but that means we would have to turn those options back into string argument form when we invoke dex2oat. Would this be less yucky if I did it that way?

Patch-set: 1