Update patch set 4
Patch Set 4:

Sorry, I postponed working on this because of the issues with the MIPS64 emulator. Now that I have a working version of the emulator I'm taking another look at the issue but I see by rereading the history that Aart was asking for a MIPS32 solution, but this patch is for MIPS64.

Aart, do you still want me to focus on MIPS32 first?

Did everyone agree on the correct algorithm for this?

For float I've found the code:

  649       public static int round(float a) {
  650           if (a != 0x1.fffffep-2f) // greatest float value less than 0.5
  651               return (int)floor(a + 0.5f);
  652           else
  653               return 0;
  654       }

There is similar code for double.

Patch-set: 4