Update patch set 1
Patch Set 1:

> (1 comment)

Another fix is more on the line of the 2nd approach in my bug update. But instead of truncating the block, we can simply replace a field/array get on null constant with constant 0/constant null, depending on the primitive type of the instruction. The proceeding null check is going to throw anyway. Need to do it in both gvn and lse. That might be a better solution?

Patch-set: 1
Label: Presubmit-Verified=0