Update patch set 2
Patch Set 2: Code-Review+1

I think the bug is critical because the StackMaps can be parsed at runtime while searching through. I know for sure that the oatdump crashes on the broken StackMaps, as it reads and decodes all of them.
Another solution would be to clear the environment when we clear NeedsEnvironment, but doing so we would loose useful debug/profiling info.

Patch-set: 2
Label: Code-Review=+1