Update patch set 2
Patch Set 2:

PS2 addresses the case of non-direct recursive inheritance.  I moved the checks from the class linker to the Dex file verifier, as it seemed more natural and simpler.

The regression test is gone in this patch set (as I could not use Smali anymore), but I'll reimplement and extend it in a follow-up PS.

Patch-set: 2