Update patch set 2
Patch Set 2:

> for (i : interesting_set_of_values) {
 > __ LoadConstant64(i);
 > expected.append(generate what you expect here);
 > }
 > DriverStr(expected, "AartLabel");
 > 
 > but with carefully chosen utilities, that can still be useful and
 > is definitely more readable than all those tests with hard coded
 > values (and did you generate those somehow ? :-)

Those 1K tests were generated by using LoadConst64() or a variant of it. But the thing is, effectively comparing LoadConst64()'s output with LoadConst64()'s output doesn't tell us anything about LoadConst64()'s correctness. Manually validating those 1K cases is impractical. We either need fewer cases or we need this automated. And if it's automated, there needs to be proper validation, without which we enter the realm of beliefs. If we believe LoadConst64() is correct, comparing LoadConst64() with LoadConst64() won't anyhow reinforce or support the belief. Further, if we believe it's correct, then why do we even try to test it? Are we having doubts about our beliefs? :) So, about the only benefit of duplicating LoadConst64() in the test is to detect accidental changes in the original LoadConst64(). And LoadConst64() isn't something that people change often or something that can silently break by itself (we test the underlying instructions).

 > perhaps if you
 > work "the other way" (start with some interesting *sub*operands
 > inside an expected pattern and construct the interesting value from
 > that) you can avoid the feeling you have to duplicate too much
 > logic in the test and still test something useful?

The 1K test cases were generated to reach each of those if/else paths in LoadConst64(). Perhaps, we got carried away and generated too many of them. We can take a look to see if those can be reduced so that the remaining cases are more practical (in terms of their number) for human validation. But if we want to validate more cases, we really want to have a way to do such validation properly and not by validating LoadConst64() with its twin.

Am I confused or being confusing?

Patch-set: 2