Update patch set 1
Patch Set 1:

I don't understand the SetStateUnsafe use well, but it seems like it was intentional as a special case.

// Suspend all threads and exclusively acquire the mutator lock. Set the state of the thread
// to kRunnable to avoid scoped object access transitions. Remove the debugger as a listener
// and clear the object registry.

// Avoid use, callers should use SetState. Used only by SignalCatcher::HandleSigQuit, ~Thread and
// Dbg::Disconnected.

https://android.googlesource.com/platform/art/+/62d6c772205b8859f0ebf7ad105402ec4c3e2e01

Does anyone know?

With that aside, in RunCheckpoint, since the thread was suspended just above in the same function, the thread is known to be suspended at the ModifySuspendCount call site, the thread won't exit even without the thread list lock held. It is safe to skip this check in this context at least.

Patch-set: 1
Label: Verified=0