Update patch set 3
Patch Set 3:

The issue was caused by my "optimization" that caches the callee method during stack unwind so that we only do look up for the top frame. However when the method is StringFactory method and it's not the top method, no lookup is done. But  it fails the IsStringInit() test since it's rewritten and not a string constructor. So I changed it back to exclusively doing the IsStringInit() test based on lookup from the caller.

Patch-set: 3
Label: Verified=0