Update patch set 2
Patch Set 2:

> Try running `vogar --benchmark $ANDROID_BUILD_TOP/libcore/benchmarks/src/benchmarks/InvokeInterface.java` both with and without this change.

Yesterday I was not able to do it quickly.
I agree with you that this change introduce regression in performance but I did not expect that it is so big because we added only one load.
We will try to reduce performance regression. Also we will check that IMT fills correctly (thank you for advice).

We know that some tests fails on this patch (for example 055-enum-performance)

Patch-set: 2
Label: Verified=0