Update patch set 3
Patch Set 3:

(2 comments)

I think this at least needs some comments to explain all the dubious assumptions we're making here.  Based on a conversation with bdc, I think the assumption is that we don't normally shut down cleanly, so this code is likely to be executed only in test scenarios.

But this is still going to time out regularly for the oddball app that repeatedly creates VMs in a loop, at least of it also has at least one unattached thread that doesn't go away.  I would guess that's enough of a problem that the timeout has to be a lot shorter than 2 seconds total.

The main down side of not actually unloading libart seems to be that the library doesn't get reinitialized if you try to start ART a second time.  I don't know how much of a problem this causes compared to the issues with this approach.  If it just breaks valgrind, we could sleep for a couple of seconds in just the valgrind case.

Patch-set: 3
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0