Update patch set 1
Patch Set 1:

Aart, just imaging that array access in a loop is guarded by if to access an array only if there is no out of bound condition. BCE will be smart enough to prove that array access is safe. But if you move array access to pre-header without if that may lead to the crash. So my first statement that to hoist invariant should be executed at once is incorrect actually, we should prove that it will be executed on the first iteration. Anton's patch actually a bit stronger - it checks that it is executed on each iteration while it would be enough that it is executed at least once.
Anyway let's wait a bit until Anton will upload a test.

Patch-set: 1