Update patch set 8
Patch Set 8: Verified+1

This seems to fix the falkiness as far as i can tell. Note that we need to use an observer pattern during shutdown because the vm is destroyed as well so we cannot get the result otherwise without reading freed memory.

Since the runtime is shutting down we cannot really use our normal mutexs but since this is all single threaded they are not really necessary.

Patch-set: 8
Label: Verified=+1