Update patch set 4
Patch Set 4:

> The worst thing happens then is redundant memory access for the
 > loop which does not execute at least once and this is a low
 > probability.

Yes, this is the case I was originally alluding to. If bce kicks in, and the array is not conditional in the loop, then not iterating at all should not make a correctness difference, other than loading something that is not used. Please go ahead and make the changes I suggested. Also please add the unit tests I added in the comment. That one actually uses part of the array in the result.

Thanks!

Patch-set: 4
Label: Code-Review=0