Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

As explained in the public bug, I think it is more a general issue, not only related to these thread events.

We need to investigate why the weak reference is not deleted even if the java.lang.Thread (created for the time of the callback from JNI) is garbage collected.

Patch-set: 1
Reviewer: Gerrit User 1029223 <1029223@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1