Update patch set 3
Patch Set 3:

I don't know whether the rewrite wrt/ daemons is what we want. Part of having the daemons was to check that shutdown is OK. Which is exactly not tested now, with the daemons being stopped explicitly.

Point could be made that test 132 was introduced afterwards to do some of this. But I'm still not really happy taking the workload away (e.g., 132 doesn't test a daemon doing actual work).

Could you at least make the daemons "restart" after completing?

Patch-set: 3