Update patch set 3
Patch Set 3:

Sorry for the delay, I had the comments prepared but never pressed post (mostly same as Nicolas).

To me it looks a bit unnecessary to thread profile_changed to all the methods. Seems cleaner to just pass it in the constructor as we don't have other use cases. But if it looks better to you feel free to go ahead with the refactoring.

Patch-set: 3
Label: Presubmit-Verified=0