Update patch set 2
Patch Set 2:

> what direction are you going to take with the tests? I assume this
 > CL is still active?

Yes, the test for LoadConst64() is being worked on. We're moving the body of LoadConst64() to a template, which can then be instantiated to work either with the assembler (to generate instructions) or with a small assembler mock (to test the effect of the generated instructions). We'll upload an update when we're done.

Patch-set: 2
Label: Verified=0