Update patch set 1
Patch Set 1:

I've been considering this optimization for some time but it's not that simple. This is breaking the type system. Consider

  int foo(short[] a) {
    int x = a[0] & 0xffff;  // Unsigned load?
    return (short) x;       // Implicit conversion?
  }

We would first convert the "a[0] & 0xffff" to an unsigned load but the type of that load would remain "short", so the VisitTypeConversion() would then eliminate the short-to-short conversion and the generated code would incorrectly zero-extend instead of sign-extending.

Maybe we could do this in some late pass, such as PrepareForRegisterAllocation, to avoid such interference but then we could miss out on other potential optimizations, such as merging unsigned byte loads, shifts and adds to a single unaligned load.

But I believe the best option would be to also change the type that the instruction returns. For short, we can reuse char but for byte we need to introduce the "unsigned byte" which does not exist in Java.

Patch-set: 1
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0