Update patch set 3
Patch Set 3:

> Looks too narrow. Let's discuss offline.

Talked offline. The example in the comment I've put in the code is not describing fully the optimization. The optimization does handle multiple incoming blocks into the if block.

What the optimization does not handle and could be improved in a follow-up change, is if the if block has multiple phis.

I've improved the comment.

Patch-set: 3