Update patch set 3
Patch Set 3:

> Switches in this CL allow us to disable/enable optimizations rather
 > than compilation. I think this approach is better then the
 > compiled-methods one. This way we don't assume that the interpreter
 > is correct. Also one of the features we aim to implement is finding
 > runtime regressions, which seems impossible with interpreting being
 > 10-100x slower then running compiled code.

Citation needed for 100x.

The argument is against you introducing *another* set of parameters controlling the compilation of methods, and that's what you're doing. You introduce *two* new parameters to change compiler behavior, in an IMHO super ugly way ("up to method-number" is something really horrible, because it isn't really well-defined). That *could* be rolled into the existing parameter, if that is generalized.

Patch-set: 3