Class Hierarchy Analysis (CHA)
The class linker now tracks whether a method has a single implementation
and if so, the JIT compiler will try to devirtualize a virtual call for
the method into a direct call. A guard is added which checks a hidden
should_deoptimize flag for deoptimization. The flag is set if additional
class linking invalidates the single-implementation assumption.

This CL does CHA-based devirtualization only if the method can be
inlined as a result so it's a cheaper form of inline cache. It also
does not devirtualize abstract/interface method invocation.

A guard can be eliminated under certain circumstances. This CL only
eliminates a guard for a devirtualized call whose receiver is a
parameter.

Change-Id: I18bf716a601b6413b46312e925a6ad9e4008efa4
Test: m test-art-host/target