Update patch set 2
Patch Set 2:

(3 comments)

NG: Because I was planning to add a use of Mockito to a test that is run by the buildbot:

https://android-review.googlesource.com/#/c/249662/7/luni/src/test/java/libcore/java/lang/ThreadTest.java

I can of course rewrite this particular test to not use Mockito; but I was assuming that we were quite unhappy about the fact that we cannot include any of the tests that use Mockito in the buildbot run. Was that assumption mistaken?

Patch-set: 2