Update patch set 4
Patch Set 4:

If i'm not mistaken, we had other tests in URLConnection that used mockito but you (tobias) removed them.

I agree with Paul's sentiment in general, but we need to make an explicit decision and stick with it. If we commit to writing tests with mockito, we need 0.1 or 0.05 of an engineer to make sure things continue to run.

Patch-set: 4