Update patch set 2
Patch Set 2: Code-Review+1

I'm very much in favour of this.

Are you planning to do it for `arm` as well? We are almost ready to upload the first VIXL32 patches, and we will need to update them to follow this.

Patch-set: 2
Reviewer: Gerrit User 1052304 <1052304@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1