Add basic runtime-plugins support.
This allows one to pass shared-libraries on the command line that the
runtime will load as plugins. They have access to runtime code and can
install hooks to add functionality. Currently the only hook they can
touch is JavaVMExt::AddEnvironmentHook to register a callback for
GetEnv(). More hooks might be added in the future.

Note that the test does not currently run on target due to not syncing
all the files.

Change-Id: I852b4daf5a3fa71e9888722bc07794632c0e5010
Test: ./test/run-test --host 900