Bisection bug search script
Bisection Bug Search is a tool for finding compiler optimization
bugs. It accepts a program which exposes a bug by producing incorrect
output and expected correct output for the program. The tool will
then attempt to narrow down the issue to a single method and
optimization pass.

More specifically the tool tries to find such method M and pass P that
when compiling only method M with optimizations up to but _excluding_
pass P and interpreting all other methods program produces correct
output, but when compiling only method M with optimizations up to and
_including_ pass P and interpreting all other methods program produces
incorrect output.

Test: unit tests ./art/tools/bisection-search/tests.py
Manual testing

Change-Id: Ic40a82184975d42c9a403f697995e5c9654b8e52