Bisection bug search script
Bisection Bug Search is a tool for localizing compiler optimization
bugs. It accepts a program which exposes a bug by producing incorrect
output and expected correct output for the program. The tool will
then attempt to narrow down the issue to a single method and
optimization pass.

More specifically the tool tries to find such method M and pass P that
when compiling only method M with optimizations up to but _excluding_
pass P and interpreting all other methods program produces correct
output, but when compiling only method M with optimizations up to and
_including_ pass P and interpreting all other methods program produces
incorrect output.

Test:
Manual testing, example:
1. Download:
https://drive.google.com/open?id=0B-J3UL0pHz9pN2R5T2gxVHFrelE
https://drive.google.com/open?id=0B-J3UL0pHz9pUU94RGVvenBmRUU
2. Save in aosp root as classes.dex and out_int
3. Cd to aosp root and run:
./art/tools/bisection-search.py -cp classes.dex --correct-output out_int
--host Test19742
4. Verify output and logfile for lack of errors

Change-Id: Ic40a82184975d42c9a403f697995e5c9654b8e52