Add an access() check in ImageCreationAllowed().
A new zygote process, the webview_zygote, will spawn isolated_app children,
but the process itself does not run as root and cannot write to the dalvik
cache. In order to not print SELinux avc errors, check for file-write
permission when testing to see if a cache image should be written.

Test: m test-art
Test: angler boots

Bug: 21643067
Change-Id: I9f797fe7332a21ef4c6b83d2210b1673af09de1b