Add -Xzygote-unprivileged option for a zygote process that does not run as root.
The webview_zygote is a new, independent zygote process that creates isolated_app
children, but the process itself does not run as root. This flag is used to avoid
performing operations that require privileges only the app_process zygote has.

This also renames the -Xzygote flag to -Xzygote-privileged to explicitly
differentiate it from the unprivileged one.

Test: m test-art
Test: angler boots

Bug: 21643067
Change-Id: I9f797fe7332a21ef4c6b83d2210b1673af09de1b