Update patch set 1
Patch Set 1:

This originated from finding out that the JNI tests are run using the ARM assembler. We thought that with this patch we at least make sure that we are not using the ARM Assembler (when we should not).

We can remove it completely now. We were planning on doing when the VIXL backends are up and running. So would you like us to remove the ARM assembler?

Are there any other non-obvious uses of the ARM Assembler? 

Thanks!

Patch-set: 1