Remove bitrotted ART_TARGET_CODEGEN_ARCHS options
ART_TARGET_CODEGEN_ARCHS was supposed to allow building only the
compiler architectures used on the device, reducing the size of
libart-compiler, but the option was never used so it bitrotted.

Remove it for now in preparation for converting parts of art to Soong.
Hopefully all of art will be converted by the time the feature is
re-enabled, so the new implementation can be in Soong.

Bug: 30928847
Test: mmma -j art
Change-Id: I70d8dbd389a409f80f4db46a96b463f7c32d0e05