Update patch set 1
Patch Set 1:

(13 comments)

It was my suggestion because we have a gaping hole for on-device logging right now.

It just goes to logcat and that's it. People have to manually wade through the logcat to see what's going on (and hopefully you didn't step away from your desk while doing a long test run because then logcat buffer will disappear). Also if it calls out to other processes it's even harder to tell what's going on since the PIDs are different. So that makes it unscriptable as-is.

--------------

My suggestion was to have a on-device logging redirection that forwards the settings to dex2oat and patchoat which would also make it extremely convenient to capture any logs for a *specific* run without all the other junk that might happen in the background.

This is similar to how we can turn on verbosity per-component with -verbose:jni etc, so I think it's fairly consistent with with our logging methodology.

Patch-set: 1
Label: Presubmit-Verified=0