Update patch set 1
Patch Set 1:

> adb logcat already supports redirecting logs to files based on severity. Why is that not sufficient? (see 'adb logcat --help')

Simultaneous runs can interleave with different PIDs. Can filter by PID but then if the runtime ends up dispatching to dex2oat or patchoat we'll be over-filtering and losing out log data.

> I remember hearing people say persisting logcat data has implications for privacy. It would be good to double check this isn't violating Android privacy policies.

Where is this coming from? It's not changing how any of the existing system integration is working (e.g. from package manager) so the persistence is exactly the same as before.

Patch-set: 1