Update patch set 1
Patch Set 1:

This feature will likely need a gate on ro.debuggable, turned off for "user" builds. I believe this will address the privacy concerns.

Why not turn on logpersist (now a Developer Option on "eng" and "userdebug" builds), then create automated tools to post-process the data in the bugreport? All of the data you are trying to deploy is there. Maybe piggy-back on b/30587474 and add the 'features' to track the interesting information there?

Patch-set: 1