Update patch set 1
Patch Set 1:

reading an environment variable is cheap, we can redirect to STDOUT or STDERR in liblog if necessary. But environment variables do not necessarily propagate to the children of a process. At issue as well is stderr and stdout may be CLOEXEC and as a result may also be closed on the children.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0