Update patch set 1
Patch Set 1:

I talked with Mark offline and we came up with a few alternatives (listed more-to-less desirable):

1) Use liblog event logging to log in the parent pid + child pid every time there's a fork (e.g. in art) -- then use logcat post-processing to figure out how to capture one pid's recursive logs.

2) Consider integrating something like (bionic's) android::base::Logging which already supports a StderrLogger/LogdLogger parameters for their InitLogging function. We'd have to integrate our VLOG(component) support somehow but it seems to do the rest of the work we already do in art and more.

3) liblog can record parent pids itself but it's not great because the parent PID can change by the time you do a log, also it's untrustworthy unless the kernel side is also changed to put the PPID as part of the credential.

------------

and of course we'd probably want to cut general support for redirecting to a file, all this really needs is redirection to stdout[/stderr] and then higher level tools could redirect to a file if they absolutely wanted to.

We may want to do #1 even if it's just for the sake of having more accurate bugreports (= no guessing which subprocesses belongs to what).

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0